[SurfaceGeometrySmoother] move functions to geometry
